Go back to DB for currentSettings instead of keeping them in memory to prepare for using multiple sets of settings.
This is slightly more expensive but will simplify code when using multiple sets of settings. Since settings are only needed when registering a new client or updating the pipeline state, this shouldn't have much of an effect since both should not be very frequest cases the way build monitors are usually used.